fix: align unified site route contracts
