+++ packages/cli/src/serve/namespace.ts
@@ -10,18 +10,18 @@
  * Bridge routes (`/__pyric/mcp`, `/__pyric/sandbox`) mount here in P2.
  */
 import { randomBytes } from 'node:crypto';
-import { existsSync, readFileSync, statSync } from 'node:fs';
-import { basename, extname, join } from 'node:path';
+import { existsSync } from 'node:fs';
+import { basename, join } from 'node:path';
 import type { IncomingMessage, ServerResponse } from 'node:http';
 import type { ActivityIncident } from 'pyric/firestore/internal';
 import { collectBody } from '../bridge/server/peer.js';
 import { StateFileError, type StateSection, type StateStore } from './state-store.js';
 import { createWriterLock, type WriterLock } from './writer-lock.js';
 import { createStudioRoutes, type StudioRouteOptions } from './studio/index.js';
-import { contentTypeFor, pipeFileToResponse, resolveStaticFile, type ServeLogger } from './server.js';
-import { SANDBOX_BUILD_MARKER } from './sandbox-marker.js';
+import { pipeFileToResponse, type ServeLogger } from './server.js';
 import type { InitPayload } from './init-payload.js';
 import { handleActivity } from './activity-route.js';
+import { createSiteTreeHandler } from './site-tree.js';
 export type { InitPayload } from './init-payload.js';
 
 /**
@@ -103,31 +103,6 @@ export interface NamespaceOptions {
    *  Absent ⇒ diagnostics are dropped (a caller that wires no logger opts out
    *  silently rather than falling back to raw `console`). */
   logger?: ServeLogger;
-}
-
-interface StudioRoutesManifest {
-  routes: string[];
-}
-
-function studioRoutesFromSite(root: string): ReadonlySet<string> {
-  const manifestPath = join(root, 'studio-routes.json');
-  const parsed = JSON.parse(readFileSync(manifestPath, 'utf8')) as StudioRoutesManifest;
-  if (!Array.isArray(parsed.routes) || parsed.routes.some((route) => typeof route !== 'string')) {
-    throw new Error(`pyric: invalid Studio route manifest at ${manifestPath}`);
-  }
-  return new Set(parsed.routes);
-}
-
-function withWorkerVersion(html: string, workerVersion: string | undefined): string {
-  if (!workerVersion || html.includes('name="pyric-worker-v"')) return html;
-  if (!/^[a-f0-9]{16}$/.test(workerVersion)) {
-    throw new Error(`pyric: invalid SharedWorker epoch '${workerVersion}'`);
-  }
-  const meta = `<meta name="pyric-worker-v" content="${workerVersion}" data-pyric-serve>`;
-  const head = html.match(/<head[^>]*>/i);
-  if (!head || head.index === undefined) return meta + html;
-  const at = head.index + head[0].length;
-  return html.slice(0, at) + meta + html.slice(at);
 }
 
 // ─── AI proxy (`/__pyric/ai-proxy` — pyric/ai, cdd-deltas #98.2) ───────────
@@ -484,7 +459,9 @@ async function handleDenials(
 export function createPyricNamespace(opts: NamespaceOptions) {
   const writerLock = createWriterLock();
   const studioRoutes = opts.studio ? createStudioRoutes(opts.studio) : null;
-  const siteStudioRoutes = opts.siteUiDir ? studioRoutesFromSite(opts.siteUiDir) : null;
+  const siteTree = opts.siteUiDir
+    ? createSiteTreeHandler(opts.siteUiDir, opts.workerVersion)
+    : null;
   const denialThrottle = createDenialThrottle();
   // Issued once per server boot. The outer static/Vite host guard protects
   // init.json before this capability is disclosed to the served runtime.
@@ -535,130 +512,7 @@ export function createPyricNamespace(opts: NamespaceOptions) {
       pipeFileToResponse(file, res);
       return true;
     }
-    if (
-      opts.siteUiDir &&
-      siteStudioRoutes &&
-      (url.pathname === '/__pyric/ui' || url.pathname.startsWith('/__pyric/ui/'))
-    ) {
-      if (url.pathname === '/__pyric/ui') {
-        res.writeHead(301, { location: '/__pyric/ui/' }).end();
-        return true;
-      }
-      const rel = url.pathname.slice('/__pyric/ui'.length) || '/';
-      if (!rel.endsWith('/') && !extname(rel)) {
-        const dir = join(opts.siteUiDir, decodeURIComponent(rel));
-        if (existsSync(dir) && statSync(dir).isDirectory()) {
-          res.writeHead(301, { location: `${url.pathname}/` }).end();
-          return true;
-        }
-      }
-
-      const first = rel.split('/').filter(Boolean)[0] ?? 'home';
-      const studioRequest = first === 'home' || siteStudioRoutes.has(first);
-      let file = resolveStaticFile(opts.siteUiDir, rel);
-      if (!file && studioRequest) {
-        const entry = first === 'home' ? '/index.html' : `/${first}/index.html`;
-        file = resolveStaticFile(opts.siteUiDir, entry);
-      }
-      if (!file) {
-        res.writeHead(404).end('not found');
-        return true;
-      }
-
-      const contentType = contentTypeFor(file);
-      res.writeHead(200, { 'content-type': contentType, 'cache-control': 'no-store' });
-      if (studioRequest && contentType.includes('text/html')) {
-        res.end(withWorkerVersion(readFileSync(file, 'utf8'), opts.workerVersion));
-      } else {
-        pipeFileToResponse(file, res);
-      }
-      return true;
-    }
+    if (siteTree?.(req, res, url)) return true;
     return false; // unknown /__pyric/* → caller 404s
   };
-}
-
-// ─── HTML injection ───────────────────────────────────────────────────
-
-/** The import-map targets. Spec → served URL. */
-export function sdkImportMap(): Record<string, string> {
-  return {
-    'firebase/ai': '/__pyric/sdk/ai.js',
-    'firebase/app': '/__pyric/sdk/app.js',
-    'firebase/auth': '/__pyric/sdk/auth.js',
-    'firebase/database': '/__pyric/sdk/database.js',
-    'firebase/firestore': '/__pyric/sdk/firestore.js',
-    'firebase/messaging': '/__pyric/sdk/messaging.js',
-    'firebase/messaging/sw': '/__pyric/sdk/messaging-sw.js',
-    'firebase/storage': '/__pyric/sdk/storage.js',
-  };
-}
-
-/**
- * Inject the import map + init script into an HTML document. The import map
- * MUST precede any module script that imports a mapped specifier, so both
- * tags go at the very start of `<head>` (fallbacks: after `<html>`, else
- * prepended). Idempotent — a page already carrying the marker is untouched
- * (matters when transformHtml runs over an SPA fallback repeatedly).
- */
-export function injectServeTags(
-  html: string,
-  importMap: Record<string, string> = sdkImportMap(),
-  workerVersion?: string,
-  forceInPage = false,
-): string {
-  const MARKER = 'data-pyric-serve';
-  if (html.includes(MARKER)) return html;
-  // A pyric SANDBOX BUILD (`vite build` under the pyricSandbox plugin's sandbox
-  // mode) already BUNDLES its own runtime + init chunk — injecting the import
-  // map + /__pyric/sdk/init.js on top would boot a SECOND runtime instance on
-  // the page (two banners, two bridge peer registrations, races between them).
-  // The bundle owns the sandbox for a marked page: it fetches
-  // /__pyric/init.json itself (worker path: the SharedWorker does; in-page
-  // path: the runtime does) and owns rules hot-reload the same way. The ONLY
-  // serve-time contribution left is the worker-version staleness stamp, which
-  // the bundled runtime reads from this meta to warn about a stale
-  // still-running SharedWorker.
-  if (html.includes(SANDBOX_BUILD_MARKER)) {
-    if (!workerVersion || html.includes('pyric-worker-v')) return html;
-    const meta = `<meta name="pyric-worker-v" content="${workerVersion}" ${MARKER}>`;
-    const headTag = html.match(/<head[^>]*>/i);
-    if (headTag && headTag.index !== undefined) {
-      const at = headTag.index + headTag[0].length;
-      return html.slice(0, at) + meta + html.slice(at);
-    }
-    return meta + html;
-  }
-  // Stamp the worker's content hash so the page can DETECT staleness: a live
-  // SharedWorker survives serve restarts + reloads and can't hot-update, so it
-  // keeps running old code until every tab of the origin closes. The worker
-  // name is intentionally STABLE (one backend shared by all tabs), so rather
-  // than split tabs across versions, `runtime.ts` compares this served hash to
-  // the running worker's baked hash and warns the user to close all tabs.
-  const versionMeta = workerVersion
-    ? `<meta name="pyric-worker-v" content="${workerVersion}" ${MARKER}>`
-    : '';
-  // `--bridge`: the MCP bridge peers with the IN-PAGE sandbox, so force the page
-  // off the default SharedWorker path. Otherwise the agent drives an empty
-  // in-page sandbox while the app's data lives in the worker. Set before any app
-  // code runs; mirrors the Vite plugin's transformIndexHtml.
-  const forceTag = forceInPage
-    ? `<script ${MARKER}>globalThis.__PYRIC_FORCE_INPAGE__=true;</script>`
-    : '';
-  const tags =
-    versionMeta +
-    forceTag +
-    `<script type="importmap" ${MARKER}>${JSON.stringify({ imports: importMap })}</script>` +
-    `<script type="module" src="/__pyric/sdk/init.js" ${MARKER}></script>`;
-  const head = html.match(/<head[^>]*>/i);
-  if (head && head.index !== undefined) {
-    const at = head.index + head[0].length;
-    return html.slice(0, at) + tags + html.slice(at);
-  }
-  const htmlTag = html.match(/<html[^>]*>/i);
-  if (htmlTag && htmlTag.index !== undefined) {
-    const at = htmlTag.index + htmlTag[0].length;
-    return html.slice(0, at) + tags + html.slice(at);
-  }
-  return tags + html;
 }
