+++ packages/cli/test/serve/namespace.test.ts
@@ -1,14 +1,11 @@
 /** `/__pyric/` namespace + HTML injection (plan step 1.4). */
 import { afterEach, describe, expect, it } from 'bun:test';
-import { mkdirSync, mkdtempSync, writeFileSync } from 'node:fs';
+import { mkdtempSync, writeFileSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
-import {
-  createPyricNamespace,
-  injectServeTags,
-  sdkImportMap,
-} from '../../src/serve/namespace.js';
-import { loopbackHosts, silentServeLogger, startStaticServer, type ServeHandle } from '../../src/serve/server.js';
+import { createPyricNamespace } from '../../src/serve/namespace.js';
+import { injectServeTags } from '../../src/serve/html-injection.js';
+import { silentServeLogger, startStaticServer, type ServeHandle } from '../../src/serve/server.js';
 
 function fixture() {
   const dir = mkdtempSync(join(tmpdir(), 'pyric-serve-ns-'));
@@ -26,84 +23,6 @@ function fixture() {
 const handles: ServeHandle[] = [];
 afterEach(async () => {
   while (handles.length) await handles.pop()!.stop();
-});
-
-describe('injectServeTags', () => {
-  it('injects import map + init script at the start of <head>, idempotently', () => {
-    const out = injectServeTags('<html><head><script type="module" src="/app.js"></script></head></html>');
-    const mapAt = out.indexOf('type="importmap"');
-    const initAt = out.indexOf('/__pyric/sdk/init.js');
-    // match the script tag, not the import map's own ".../sdk/app.js" value
-    const appAt = out.indexOf('src="/app.js"');
-    expect(mapAt).toBeGreaterThan(-1);
-    // the import map must precede the app's module script
-    expect(mapAt).toBeLessThan(appAt);
-    expect(initAt).toBeLessThan(appAt);
-    expect(out).toContain('"firebase/auth":"/__pyric/sdk/auth.js"');
-    expect(injectServeTags(out)).toBe(out); // idempotent
-  });
-
-  it('falls back when <head> is absent', () => {
-    expect(injectServeTags('<html><body>x</body></html>')).toContain('importmap');
-    expect(injectServeTags('no tags at all')).toContain('importmap');
-  });
-
-  it('stamps the worker version meta when provided (versioned SharedWorker name)', () => {
-    const out = injectServeTags('<html><head></head></html>', undefined, 'abc123');
-    expect(out).toContain('<meta name="pyric-worker-v" content="abc123"');
-    // precedes the importmap so it's parsed before the runtime reads it
-    expect(out.indexOf('pyric-worker-v')).toBeLessThan(out.indexOf('importmap'));
-    // omitted by default → no meta
-    expect(injectServeTags('<html><head></head></html>')).not.toContain('pyric-worker-v');
-  });
-
-  it('forces the in-page sandbox when forceInPage is set (worker-unavailable fallback), before init', () => {
-    const out = injectServeTags('<html><head></head></html>', undefined, undefined, true);
-    expect(out).toContain('__PYRIC_FORCE_INPAGE__=true');
-    // must be set before the init script reads it
-    expect(out.indexOf('__PYRIC_FORCE_INPAGE__')).toBeLessThan(out.indexOf('/__pyric/sdk/init.js'));
-    // default (forceInPage off) does NOT force in-page
-    expect(injectServeTags('<html><head></head></html>')).not.toContain('__PYRIC_FORCE_INPAGE__');
-  });
-
-  it('SKIPS injection for a marked sandbox build page (the bundle owns the sandbox)', () => {
-    const marked =
-      '<html><head><meta name="pyric-sandbox-build" content="1" data-pyric-sandbox-build></head><body></body></html>';
-    const out = injectServeTags(marked, undefined, 'abc123');
-    // no second runtime: neither the import map nor the injected init module
-    expect(out).not.toContain('importmap');
-    expect(out).not.toContain('/__pyric/sdk/init.js');
-    // the ONE serve-time contribution: the worker staleness stamp
-    expect(out).toContain('<meta name="pyric-worker-v" content="abc123"');
-    // idempotent
-    expect(injectServeTags(out, undefined, 'abc123')).toBe(out);
-    // without a worker version, the marked page passes through untouched
-    expect(injectServeTags(marked)).toBe(marked);
-  });
-
-  it('sdkImportMap covers the served modules', () => {
-    expect(Object.keys(sdkImportMap()).sort()).toEqual([
-      'firebase/ai',
-      'firebase/app',
-      'firebase/auth',
-      'firebase/database',
-      'firebase/firestore',
-      'firebase/messaging',
-      'firebase/messaging/sw',
-      'firebase/storage',
-    ]);
-  });
-});
-
-describe('loopbackHosts (dual-family bind — the IPv4/IPv6 trap fix)', () => {
-  it('localhost binds BOTH loopback families so 127.0.0.1 AND ::1 reach the server', () => {
-    expect(loopbackHosts('localhost')).toEqual(['127.0.0.1', '::1']);
-  });
-  it('an explicit host binds only itself (never all-interfaces — sandbox stays off the LAN)', () => {
-    expect(loopbackHosts('0.0.0.0')).toEqual(['0.0.0.0']);
-    expect(loopbackHosts('192.168.1.5')).toEqual(['192.168.1.5']);
-    expect(loopbackHosts('127.0.0.1')).toEqual(['127.0.0.1']);
-  });
 });
 
 describe('namespace over the real server', () => {
@@ -188,66 +107,6 @@ describe('namespace over the real server', () => {
     const html = await (await fetch(h.url + '/')).text();
     expect(html).toContain('type="importmap"');
     expect(html).toContain('/__pyric/sdk/init.js');
-  });
-
-  it('serves one Astro site tree with scoped Studio fallbacks and hard docs/asset 404s', async () => {
-    const { site, sdk } = fixture();
-    const siteRoot = join(site, 'apps', 'site');
-    mkdirSync(join(siteRoot, 'firestore'), { recursive: true });
-    mkdirSync(join(siteRoot, 'storage'), { recursive: true });
-    mkdirSync(join(siteRoot, 'docs', 'overview'), { recursive: true });
-    mkdirSync(join(siteRoot, '_astro'), { recursive: true });
-    writeFileSync(join(siteRoot, 'index.html'), '<!doctype html><head></head>HOME');
-    writeFileSync(
-      join(siteRoot, 'firestore', 'index.html'),
-      '<!doctype html><head></head>FIRESTORE',
-    );
-    writeFileSync(join(siteRoot, 'storage', 'index.html'), '<!doctype html><head></head>STORAGE');
-    writeFileSync(join(siteRoot, 'docs', 'overview', 'index.html'), '<!doctype html>DOCS');
-    writeFileSync(join(siteRoot, 'docs', 'overview.md'), '# Overview');
-    writeFileSync(join(siteRoot, 'docs', 'index.json'), '{"pages":[]}');
-    writeFileSync(join(siteRoot, '_astro', 'app.js'), '// app');
-    writeFileSync(
-      join(siteRoot, 'studio-routes.json'),
-      JSON.stringify({ routes: ['home', 'firestore', 'storage'] }),
-    );
-
-    const ns = createPyricNamespace({
-      sdkDir: sdk,
-      initPayload: () => ({ rules: null, rulesHash: null, bridgeUrl: null }),
-      siteUiDir: siteRoot,
-      workerVersion: '0123456789abcdef',
-    });
-    const h = await startStaticServer({
-      publicDir: site,
-      port: 0,
-      host: '127.0.0.1',
-      portScanLimit: 200,
-      logger: silentServeLogger(),
-      namespaceHandler: ns,
-    });
-    handles.push(h);
-
-    const deep = await fetch(h.url + '/__pyric/ui/firestore/users/alice');
-    const deepHtml = await deep.text();
-    expect(deep.status).toBe(200);
-    expect(deepHtml).toContain('FIRESTORE');
-    expect(deepHtml).toContain('name="pyric-worker-v" content="0123456789abcdef"');
-    expect(await (await fetch(h.url + '/__pyric/ui/storage/uploads/logo.png')).text()).toContain('STORAGE');
-
-    expect((await fetch(h.url + '/__pyric/ui/_astro/app.js')).status).toBe(200);
-    expect((await fetch(h.url + '/__pyric/ui/_astro/missing.js')).status).toBe(404);
-    expect((await fetch(h.url + '/__pyric/ui/docs/missing')).status).toBe(404);
-    expect((await fetch(h.url + '/__pyric/ui/not-a-service')).status).toBe(404);
-
-    const docsTwin = await fetch(h.url + '/__pyric/ui/docs/overview.md');
-    expect(docsTwin.status).toBe(200);
-    expect(await docsTwin.text()).toContain('# Overview');
-    expect((await fetch(h.url + '/__pyric/ui/docs/index.json')).headers.get('content-type')).toContain('application/json');
-
-    const docs = await (await fetch(h.url + '/__pyric/ui/docs/overview/')).text();
-    expect(docs).toContain('DOCS');
-    expect(docs).not.toContain('pyric-worker-v');
   });
 
   it('GET /__pyric/capture returns the fixture (200) or 404 when absent; POST writes it', async () => {
