+++ packages/site-docs/package.json
@@ -21,10 +21,12 @@
     "tailwindcss": "^4.0.0"
   },
   "devDependencies": {
+    "@testing-library/react": "^16.1.0",
     "@types/node": "^22.0.0",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "github-slugger": "^2.0.0",
+    "jsdom": "^29.1.1",
     "playwright": "^1.60.0",
     "typescript": "^5.7.0"
   }
