+++ packages/site-docs/src/examples/registry.ts
@@ -1,5 +1,3 @@
-import firestoreFirstWrite from './firestore-first-write';
-import firestoreFirstWriteSource from './firestore-first-write/run.ts?raw';
 import type { PyricExampleDefinition } from './definition';
 
 export interface RegisteredPyricExample {
@@ -7,17 +5,32 @@ export interface RegisteredPyricExample {
   source: string;
 }
 
-export const PYRIC_EXAMPLES = {
-  'firestore-first-write': {
-    definition: firestoreFirstWrite,
-    source: firestoreFirstWriteSource,
-  },
-} as const satisfies Record<string, RegisteredPyricExample>;
+const definitions = import.meta.glob<PyricExampleDefinition>('./*/definition.ts', {
+  eager: true,
+  import: 'default',
+});
+const sources = import.meta.glob<string>('./*/run.ts', {
+  eager: true,
+  import: 'default',
+  query: '?raw',
+});
 
-export type PyricExampleId = keyof typeof PYRIC_EXAMPLES;
+const entries = Object.entries(definitions).map(([path, definition]) => {
+  const match = path.match(/^\.\/([^/]+)\/definition\.ts$/);
+  if (!match) throw new Error(`Invalid Pyric example definition path: ${path}`);
+  const id = match[1]!;
+  const source = sources[`./${id}/run.ts`];
+  if (source === undefined) throw new Error(`Pyric example '${id}' has no run.ts source`);
+  return [id, { definition, source }] as const;
+});
+
+export const PYRIC_EXAMPLES: Readonly<Record<string, RegisteredPyricExample>> =
+  Object.freeze(Object.fromEntries(entries));
+
+export type PyricExampleId = string;
 
 export function pyricExample(id: string): RegisteredPyricExample {
-  const example = PYRIC_EXAMPLES[id as PyricExampleId];
+  const example = PYRIC_EXAMPLES[id];
   if (!example) throw new Error(`Unknown Pyric example: ${id}`);
   return example;
 }
